refactor: 💡 constants 파일 위치 이동 및 import 경로 수정
