+++ frontend/src/components/Player/Player.tsx
@@ -1,13 +1,15 @@
-import { useSocketStore } from '@/stores';
 import { css } from '@emotion/react';
 import { useRef, useState } from 'react';
-import { divideSize, multiplySize } from '@/utils';
 import ReactPlayer from 'react-player';
+
+import { YOUTUBE_ERROR_MESSAGES } from '@/constants/youtube';
 import { useFraction, useToast } from '@/hooks';
+import { useSocketStore } from '@/stores';
 import { InterestYoutubeResponse, PlayerState, YoutubeRequestType } from '@/types';
-import { YOUTUBE_ERROR_MESSAGES } from '@/constants';
-import SharerDraggingIndicator from './SharerDraggingIndicator';
+import { divideSize, multiplySize } from '@/utils';
+
 import ControllerSection from './ControllerSection';
+import SharerDraggingIndicator from './SharerDraggingIndicator';
 import StateChangeIndicator from './StateChangeIndicator';
 import VideoOverlayToggle from './VideoOverlayToggle';
